feat: refactoring wallet vue application structure
- refactored wallet web application structure
- closes #814
