+++ cmd/wallet-web/src/pages/BlockNoAuth.vue
