+++ cmd/wallet-web/src/pages/RegisterDevice.vue
