+++ cmd/wallet-web/src/pages/Worker.vue
