+++ cmd/wallet-web/src/pages/demos/IssueCredential.vue
