+++ cmd/wallet-web/src/pages/demos/WebWallet.vue
