+++ cmd/wallet-web/src/pages/demos/webWalletSamples.js
