+++ cmd/wallet-web/src/pages/mixins/common/helper.js
