+++ cmd/wallet-web/src/pages/mixins/common/register.js
