+++ cmd/wallet-web/src/pages/mixins/get/didConn.js
