+++ cmd/wallet-web/src/pages/mixins/get/getCredentialsByQuery.js
