+++ cmd/wallet-web/src/pages/mixins/index.js
